Log Error Checking
If a Code Fails to Add to your save and an Error code is given from Apollo the Code will be marked with an X and also the Error will show Red in the log.
Note: It Does remove the Error codes from the Current Built Collector Patch file but does not remove them from the Screen untill u clear it.